refactor(service): move temporal/utils.ts to lib/utils.ts
Not temporal-specific — deepEqual, retryPolicy, and constants belong in
lib alongside other shared service utilities.

Committed-By-Agent: claude
