+++ apps/service/src/lib/utils.ts
